audits, questions and answers created

+++ app/controllers/UsersController.php
@@ -4,29 +4,43 @@ class UsersController extends Controller {
 
     public function index()
     {
-        $ints = new Users($this->d1);
-        $this->f3->set('users',$ints->all($this->f3->get('SESSION.company')));
-        $this->f3->set('section','users');
-        $this->f3->set('columns','[1,3,4,5]');
-        $this->f3->set('subnav','true');
-        $this->f3->set('page_head','List');
-        $this->f3->set('view','users/list.htm');
+        if ($this->f3->get('SESSION.roles') == 'Admin'){
+            $ints = new Users($this->d1);
+            $this->f3->set('users',$ints->all($this->f3->get('SESSION.company')));
+            $this->f3->set('section','users');
+            $this->f3->set('columns','[1,3,4,5]');
+            $this->f3->set('subnav','true');
+            $this->f3->set('page_head','List');
+            $this->f3->set('view','users/list.htm');
+        } else {
+            $this->f3->reroute('/');
+        }
+        
     }
     public function create()
     {
-        if($this->f3->exists('POST.create'))
-        {
-            $ints = new Users($this->d1);
-	        $hash = password_hash($this->f3->get('POST.password'),PASSWORD_DEFAULT);
-	        $this->f3->set('POST.password',$hash);
-            $ints->add();
-            $this->f3->reroute('/users');
-        }
-        else
-        {
-            $this->f3->set('mode','create');
-            $this->f3->set('page_head','New');
-            $this->f3->set('view','/users/create.htm');
+        if ($this->f3->get('SESSION.roles') == 'Admin'){
+            if($this->f3->exists('POST.create'))
+            {
+                $ints = new Users($this->d1);
+                $hash = password_hash($this->f3->get('POST.password'),PASSWORD_DEFAULT);
+                $this->f3->set('POST.password',$hash);
+                $ints->add();
+                $this->f3->reroute('/users');
+            }
+            else
+            {
+                $this->f3->set('mode','create');
+                $this->f3->set('subnav','true');
+                $this->f3->set('back','yes');
+                $this->f3->set('backto','users');
+                $this->f3->set('create','no');
+                $this->f3->set('search','no');    
+                $this->f3->set('page_head','New');
+                $this->f3->set('view','/users/create.htm');
+            }
+        } else {
+            $this->f3->reroute('/');
         }
     }
 
@@ -36,21 +50,21 @@ class UsersController extends Controller {
 
         if($this->f3->exists('POST.change'))
         {
-	    $data = $ints->getPass($this->f3->get('POST.id'));
+	        $data = $ints->getPass($this->f3->get('POST.id'));
             $hsh = $data[0]['password'];
-	    $old = $this->f3->get('POST.pass0');
+	        $old = $this->f3->get('POST.pass0');
             if (password_verify($old,$hsh)) {
-		$hash = password_hash($this->f3->get('POST.pass2'),PASSWORD_DEFAULT);
-		$this->f3->set('POST.password',$hash);
+		        $hash = password_hash($this->f3->get('POST.pass2'),PASSWORD_DEFAULT);
+		        $this->f3->set('POST.password',$hash);
             	$ints->edit($this->f3->get('POST.id'));
-		$this->f3->set('stat','success');
-		$this->f3->set('msg','Password! changed');
+		        $this->f3->set('stat','success');
+		        $this->f3->set('msg','Password! changed');
             	$this->f3->set('view','main.htm');
-	    } else {
-		$this->f3->set('stat','danger');
-		$this->f3->set('msg','Wrong Password!');
-            	$this->f3->set('view','main.htm');
-	    }
+	        } else {
+                $this->f3->set('stat','danger');
+                $this->f3->set('msg','Wrong Password!');
+                $this->f3->set('view','main.htm');
+	        }
         }
         else
         {
@@ -75,6 +89,11 @@ class UsersController extends Controller {
 	        $hash = password_hash($this->f3->get('POST.password'),PASSWORD_DEFAULT);
             $this->f3->set('hash',$hash);
             $this->f3->set('mode','update');
+            $this->f3->set('subnav','true');
+            $this->f3->set('back','yes');
+            $this->f3->set('backto','users');
+            $this->f3->set('create','no');
+            $this->f3->set('search','no');
             $this->f3->set('page_head','Update');
             $this->f3->set('view','users/update.htm');
         }
@@ -82,13 +101,16 @@ class UsersController extends Controller {
 
     public function delete()
     {
-        if($this->f3->exists('PARAMS.id'))
-        {
-            $ints = new Users($this->d1);
-            $ints->delete($this->f3->get('PARAMS.id'));
+        if ($this->f3->get('SESSION.roles') == 'Admin'){
+            if($this->f3->exists('PARAMS.id'))
+            {
+                $ints = new Users($this->d1);
+                $ints->delete($this->f3->get('PARAMS.id'));
+            }
+            $this->f3->reroute('/users');
+        } else {
+            $this->f3->reroute('/');
         }
-
-        $this->f3->reroute('/users');
     }
 
 } // main class
